refactor(leaderboard): Organize to features structure
