--- apps/leaderboard/src/components/LeaderboardTable/BadgesCell/index.ts
@@ -1,2 +1,0 @@
-export * from './badges';
-export { default as BadgesCell } from './BadgesCell';
