--- apps/leaderboard/src/components/LeaderboardTable/index.ts
@@ -1,1 +1,0 @@
-export * from './LeaderboardTable';
