+++ apps/leaderboard/src/constants/query.ts
@@ -1,0 +1,3 @@
+export const INDEXING_PROGRESS_QUERY_KEY = 'indexingProgress';
+export const LEADERBOARD_QUERY_KEY = 'leaderboard';
+export const LATEST_FINALIZED_BLOCK_QUERY_KEY = 'latestFinalizedBlock';
