+++ apps/leaderboard/src/features/indexingProgress/components/index.ts
@@ -1,0 +1,1 @@
+export * from './IndexingProgressCard';
