+++ apps/leaderboard/src/features/indexingProgress/queries/index.ts
@@ -1,0 +1,1 @@
+export * from './indexingProgressQuery';
