+++ apps/leaderboard/src/features/leaderboard/components/BadgesCell/index.ts
@@ -1,0 +1,2 @@
+export * from './badges';
+export * from './BadgesCell';
