+++ apps/leaderboard/src/features/leaderboard/components/index.ts
@@ -1,0 +1,3 @@
+export * from './BadgesCell';
+export * from './HeaderCell';
+export * from './LeaderboardTable';
