+++ apps/leaderboard/src/features/leaderboard/index.ts
@@ -1,0 +1,3 @@
+export * from './components';
+export * from './queries';
+export * from './types';
