+++ apps/leaderboard/src/features/leaderboard/queries/index.ts
@@ -1,0 +1,1 @@
+export * from './leaderboardQuery';
