+++ apps/leaderboard/src/features/leaderboard/queries/leaderboardQuery.ts
@@ -2,6 +2,7 @@ import { graphql } from '@tangle-network/tangle-shared-ui/graphql';
 import { AccountsOrderBy } from '@tangle-network/tangle-shared-ui/graphql/graphql';
 import { executeGraphQL } from '@tangle-network/tangle-shared-ui/utils/executeGraphQL';
 import { useQuery } from '@tanstack/react-query';
+import { LEADERBOARD_QUERY_KEY } from '../../../constants/query';
 
 export const TEAM_ACCOUNTS = [
   '5CJFrNyjRahyb7kcn8HH3LPJRaZf2aq6jguk5kx5V5Aa6rXh',
@@ -10,10 +11,6 @@ export const TEAM_ACCOUNTS = [
   '5FjXDSpyiLbST4PpYzX399vymhHYhxKCP8BNhLBEmLfrUYNv',
   '5Dqf9U5dgQ9GLqdfaxXGjpZf9af1sCV8UrnpRgqJPbe3wCwX',
 ] as const;
-
-export enum ReactQueryKey {
-  Leaderboard = 'Leaderboard',
-}
 
 const LeaderboardQueryDocument = graphql(/* GraphQL */ `
   query LeaderboardTableDocument(
@@ -122,7 +119,7 @@ export function useLeaderboard(
 ) {
   return useQuery({
     queryKey: [
-      ReactQueryKey.Leaderboard,
+      LEADERBOARD_QUERY_KEY,
       first,
       offset,
       blockNumberSevenDaysAgo,
