+++ apps/leaderboard/src/features/leaderboard/types/index.ts
@@ -1,4 +1,4 @@
-import { BadgeEnum } from './BadgesCell';
+import { BadgeEnum } from '../components';
 
 export interface PointsBreakdown {
   mainnet: bigint;
