+++ apps/leaderboard/src/queries/index.ts
@@ -1,0 +1,1 @@
+export * from './latestFinalizedBlockQuery';
